db initialisation + forum model changed
- changed the structure of the database initialisation
- app.js does not need to preload database connections
- forum model updated to require new db init file
